add get user by id endpoint

+++ Milestone/src/main/java/com/gcu/data/DataAccessInterface.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface DataAccessInterface <T> {
     public List<T> findAll();
-    public T FindById(int id);
+    public T FindById(long id);
     public boolean create(T t);
     public boolean update(T t);
     public boolean delete(T t);
